Bug fix to CCP4ReportParser.py (#29)
CCP4ReportParser.py in main branch included bug at line 2321 where `findall` was called on an ET.Element with an absolute instead of a relative path.  This commit replaces `//table` with `.//table` as it should be.  This codepath is not called routinely, but this fix is needed to keep/make the code suitable for migration